Reinforcement Learning/project_2/requirements.txt
box2d-py==2.3.5
Brotli
certifi
charset-normalizer
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
gymnasium==1.1.1
idna
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pygame==2.6.1
pyparsing==3.2.3
PySocks
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML
requests
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
typing_extensions
tzdata==2025.2
urllib3
win-inet-pton
